refactor(bolt-boost): refactor modules
