+++ bolt-boost/README.md
@@ -1,3 +1,4 @@
 # Bolt-Boost
 
 Bolt-Boost is a [commit-boost](https://commit-boost.github.io/commit-boost-client) module that extends the default PBS module with the [constraints-API](https://chainbound.github.io/bolt-docs/api/builder).
+It inherits the PBS module configuration for the modified `get_header` call.
